moving scope items into their own file
