+++ src/inspection/scope/index.ts
@@ -4,3 +4,4 @@
 export * from "./identifier";
 export * from "./invokeExpression";
 export * from "./scope";
+export * from "./scopeItem";
